feat: code refactoring
