+++ fare/records/config.py
@@ -16,7 +16,8 @@ from invenio_records_rest.facets import terms_filter
 from invenio_records_rest.utils import allow_all, check_elasticsearch, deny_all
 
 from ..file_management.api import FILE_MNGT_PID_FETCHER, \
-    FILE_MNGT_PID_MINTER, FILE_MNGT_PID_TYPE, RevisionSearch
+    FILE_MNGT_PID_MINTER, FILE_MNGT_PID_TYPE
+from ..file_management.search import RevisionSearch
 from .search import RevisionedRecordsSearch
